Car Ownership
Code almost finished except minor things. The program has classes for
person, driver, car and can store the information about all that but I
have not finished implementing that part. The part that is finished is
that the program also reads in a text file containing make and model of
car and stores that information in a linked list. Eventually I would
like to connect a car in the linked list being owned by a driver.
Basically in this project I try to use the concepts of aggregate
classes, inheritance, polymorphism, friend classes, overloading
operators and linked list.